pyproject.toml
[project]
name = "healthcare-ml-project"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.0",
    "joblib>=1.3.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "apache-airflow>=3.0.0",
    "kagglehub>=0.2.0",
    "asyncpg",
    "apache-airflow-providers-fab",
    "apache-airflow-providers-postgres",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["healthcare_ml_project"]